requirements.txt
requests
google-play-scraper==1.2.7
python-dotenv
msal
jinja2
playwright
